% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{as.km}
\alias{as.km}
\title{Coerce an Object into a \code{km} Object}
\usage{
as.km(x, ...)
}
\arguments{
\item{x}{Object to be coerced.}

\item{...}{Further arguments for methods.}
}
\value{
An object with S4 class \code{"km"}.
}
\description{
Coerce an object into an object with S4 class \code{"km"} from the
\pkg{DiceKriging} package.
}
\details{
Such a coercion is typically used to compare the performance of
the methods implemented in the current \pkg{rlibkriging} package to
those which are available in the \pkg{DiceKriging} package.
}
