% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asDiceKriging.R
\name{as_km.default}
\alias{as_km.default}
\title{Build a DiceKriging "km" like object.}
\usage{
\method{as_km}{default}(
  formula = ~1,
  design,
  response,
  covtype = "matern5_2",
  coef.cov = NULL,
  coef.var = NULL,
  coef.trend = NULL,
  estim.method = "MLE",
  optim.method = "BFGS",
  parinit = NULL,
  ...
)
}
\arguments{
\item{formula}{R formula object to setup the linear trend (aka Universal Kriging). Supports ~1, ~. and ~.^2}

\item{design}{data.frame of design of experiments}

\item{response}{array of output values}

\item{covtype}{covariance structure. Supports "gauss", "exp", ...}

\item{coef.cov}{fixed covariance range value (so will not optimize if given)}

\item{coef.var}{fixed variance value (so will not estimate if given)}

\item{coef.trend}{fixed trend value (so will not estimate if given)}

\item{estim.method}{estimation criterion. Supports "MLE" or "LOO"}

\item{optim.method}{optimization algorithm used on estim.method objective. Supports "BFGS"}

\item{parinit}{initial values of covariance range which will be optimzed using optim.method}

\item{...}{Ignored}
}
\value{
as_km object, extends DiceKriging::km (plus contains a "Kriging" field which contains original object)
}
\description{
Build a DiceKriging "km" like object.
}
\examples{
# a 16-points factorial design, and the corresponding response
d <- 2; n <- 16
design.fact <- expand.grid(x1=seq(0,1,length=4), x2=seq(0,1,length=4))
y <- apply(design.fact, 1, DiceKriging::branin) 

#library(DiceKriging)
# kriging model 1 : matern5_2 covariance structure, no trend, no nugget effect
#m1 <- km(design=design.fact, response=y,covtype = "gauss",parinit = c(.5,1),control = list(trace=F))
as_m1 <- as_km(design=design.fact, response=y,covtype = "gauss",parinit = c(.5,1))
}
\author{
Yann Richet (yann.richet@irsn.fr)
}
