% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asDiceKriging.R
\name{as_km.Kriging}
\alias{as_km.Kriging}
\alias{as_km,Kriging,Kriging-method}
\title{Convert a "Kriging" object to a DiceKriging::km one.}
\usage{
\method{as_km}{Kriging}(k, .call = NULL)
}
\arguments{
\item{k}{"Kriging" object}

\item{.call}{Force the "call" filed in km object}
}
\value{
as_km object, extends DiceKriging::km plus contains "Kriging" field
}
\description{
Convert a "Kriging" object to a DiceKriging::km one.
}
\examples{
f = function(x) 1-1/2*(sin(12*x)/(1+x)+2*cos(7*x)*x^5+0.7)
set.seed(123)
X <- as.matrix(runif(5))
y <- f(X)
r <- Kriging(y, X, "gauss")
print(r)
k <- as_km(r)
print(k)
}
\author{
Yann Richet (yann.richet@irsn.fr)
}
