% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asDiceKriging.R
\name{as_km}
\alias{as_km}
\title{Build a "as_km" object, which extends DiceKriging::km S4 class.}
\usage{
as_km(...)
}
\arguments{
\item{...}{args}
}
\value{
as_km/km object
}
\description{
Build a "as_km" object, which extends DiceKriging::km S4 class.
}
\author{
Yann Richet (yann.richet@irsn.fr)
}
