% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bench_solvetri}
\alias{bench_solvetri}
\title{Solve a Triangular Linear System}
\usage{
bench_solvetri(n, X, y)
}
\arguments{
\item{n}{Integer size of the system.}

\item{X}{Numeric Matrix XXXY Lower of Upper tirangular?}

\item{y}{Numeric vector: right hand side.}
}
\description{
Solve a Triangular Linear System
}
