% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrigingClass.R
\name{classKriging}
\alias{classKriging}
\title{Shortcut to provide functions to the S3 class "Kriging"}
\usage{
classKriging(nk)
}
\arguments{
\item{nk}{A pointer to a C++ object of class "Kriging"}
}
\value{
An object of class "Kriging" with methods to access and manipulate the data
}
\description{
Shortcut to provide functions to the S3 class "Kriging"
}
