% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoiseKrigingClass.R
\name{classNoiseKriging}
\alias{classNoiseKriging}
\title{Shortcut to provide functions to the S3 class "NoiseKriging"}
\usage{
classNoiseKriging(nk)
}
\arguments{
\item{nk}{A pointer to a C++ object of class "NoiseKriging"}
}
\value{
An object of class "NoiseKriging" with methods to access and manipulate the data
}
\description{
Shortcut to provide functions to the S3 class "NoiseKriging"
}
