% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NuggetKrigingClass.R
\name{classNuggetKriging}
\alias{classNuggetKriging}
\title{Shortcut to provide functions to the S3 class "NuggetKriging"}
\usage{
classNuggetKriging(nk)
}
\arguments{
\item{nk}{A pointer to a C++ object of class "NuggetKriging"}
}
\value{
An object of class "NuggetKriging" with methods to access and manipulate the data
}
\description{
Shortcut to provide functions to the S3 class "NuggetKriging"
}
