% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{copy}
\alias{copy}
\title{Duplicate object.}
\usage{
copy(object, ...)
}
\arguments{
\item{object}{An object representing a fitted model.}

\item{...}{Ignored.}
}
\value{
The copied object.
}
\description{
Duplicate a model given in
\code{object}.
}
