% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{covMat}
\alias{covMat}
\title{covariance function}
\usage{
covMat(object, ...)
}
\arguments{
\item{object}{An object representing a fitted model.}

\item{...}{Further arguments of function (eg. points, range).}
}
\value{
The covariance matrix.
}
\description{
Compute the covariance matrix of a model given in \code{object},
between given set of points.
}
