% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{fit}
\alias{fit}
\title{Fit model on data.}
\usage{
fit(object, ...)
}
\arguments{
\item{object}{An object representing a fitted model.}

\item{...}{Further arguments of function}
}
\value{
No return value. Kriging object argument is modified.
}
\description{
Fit a model given in
\code{object}.
}
