% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{get_num_nugget}
\alias{get_num_nugget}
\title{Get numerical nugget used in Cholesky decompositions}
\usage{
get_num_nugget(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
numerical nugget used.
}
\description{
Get numerical nugget used in Cholesky decompositions
}
\author{
Yann Richet \email{yann.richet@irsn.fr}
}
