% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asDiceKriging.R
\docType{class}
\name{KM-class}
\alias{KM-class}
\title{S4 class for Kriging Models Extending the \code{"km"} Class}
\description{
This class is intended to be used either by using its
    own dedicated S4 methods or by using the S4 methods inherited
    from the \code{"km"} class of the \pkg{libKriging} package.
}
\section{Slots}{

\describe{
\item{\code{d,n,X,y,p,F}}{Number of (numeric) inputs, number of
observations, design matrix, response vector, number of trend
variables, trend matrix.}

\item{\code{trend.formula,trend.coef}}{Formula used for the trend, vector
\eqn{\hat{\boldsymbol{\beta}}}{betaHat} of estimated (or fixed)
trend coefficients with length \eqn{p}.}

\item{\code{covariance}}{A}
}}

\section{Useful material}{

}

\seealso{
\code{\link[DiceKriging]{km-class}} in the
    \bold{DiceKriging} package. The creator \code{\link{KM}}.
}
\author{
Yann Richet \email{yann.richet@irsn.fr}
}
