% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{leaveOneOut}
\alias{leaveOneOut}
\title{Compute Leave-One-Out}
\usage{
leaveOneOut(object, ...)
}
\arguments{
\item{object}{An object representing a fitted model.}

\item{...}{Ignored.}
}
\value{
The Leave-One-Out sum of squares.
}
\description{
Compute the leave-One-Out error of a model given in \code{object}.
}
