% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{leaveOneOutVec}
\alias{leaveOneOutVec}
\title{Leave-One-Out vector}
\usage{
leaveOneOutVec(object, ...)
}
\arguments{
\item{object}{An object representing a fitted model.}

\item{...}{Further arguments of function (eg. range).}
}
\value{
The Leave-One-Out errors (mean and stdev) for each conditional point.
}
\description{
Compute the leave-One-Out vector error of a model given in \code{object},
at a different value of the parameters.
}
