% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoiseKrigingClass.R
\name{logLikelihood.NoiseKriging}
\alias{logLikelihood.NoiseKriging}
\alias{logLikelihood,NoiseKriging,NoiseKriging-method}
\title{Get logLikelihood of NoiseKriging Model}
\usage{
\method{logLikelihood}{NoiseKriging}(object, ...)
}
\arguments{
\item{object}{An S3 NoiseKriging object.}

\item{...}{Not used.}
}
\value{
The logLikelihood computed for fitted
    \eqn{\boldsymbol{theta}}{\theta}.
}
\description{
Get logLikelihood of NoiseKriging Model
}
\examples{
f <- function(x) 1 - 1 / 2 * (sin(12 * x) / (1 + x) + 2 * cos(7 * x) * x^5 + 0.7)
set.seed(123)
X <- as.matrix(runif(5))
y <- f(X) + 0.01*rnorm(nrow(X))
r <- NoiseKriging(y, rep(0.01^2,5), X, kernel = "gauss")
print(r)
logLikelihood(r)

}
\author{
Yann Richet \email{yann.richet@irsn.fr}
}
