% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrigingClass.R, R/allGenerics.R
\name{logLikelihood}
\alias{logLikelihood}
\title{Compute Model log-Likelihood at Given Args}
\usage{
logLikelihood(object, ...)

logLikelihood(object, ...)
}
\arguments{
\item{object}{An object representing a fitted model.}

\item{...}{Further arguments for methods.}
}
\value{
log-Likelihood

The log-likelihood.
}
\description{
Compute model log-Likelihood at given args
}
