% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{logLikelihoodFun}
\alias{logLikelihoodFun}
\title{Log-Likelihood function}
\usage{
logLikelihoodFun(object, ...)
}
\arguments{
\item{object}{An object representing a fitted model.}

\item{...}{Further arguments of function (eg. range).}
}
\value{
The log-likelihood.
}
\description{
Compute the log-Likelihood of a model given in \code{object},
at a different value of the parameters.
}
