% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrigingClass.R, R/allGenerics.R
\name{logMargPost}
\alias{logMargPost}
\title{Compute model log-Marginal-Posterior at given args}
\usage{
logMargPost(object, ...)

logMargPost(object, ...)
}
\arguments{
\item{object}{An object representing a fitted model.}

\item{...}{Further arguments for methods.}
}
\value{
log-Marginal-Posterior
 @export

The log-marginal posterior.
}
\description{
Compute model log-Marginal Posterior at given args
}
