% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{logMargPostFun}
\alias{logMargPostFun}
\title{log-Marginal Posterior function}
\usage{
logMargPostFun(object, ...)
}
\arguments{
\item{object}{An object representing a fitted model.}

\item{...}{Further arguments of function (eg. range).}
}
\value{
The log-marginal posterior.
}
\description{
Compute the log-Marginal Posterior of a model given in
\code{object}, at a different value of the parameters.
}
