% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asDiceKriging.R
\name{predict.KM}
\alias{predict.KM}
\alias{predict,KM,KM-method}
\title{Overload DiceKriging::predict.km for \code{KM} objects (expected faster).}
\usage{
\method{predict}{KM}(
  object,
  newdata,
  type = "UK",
  se.compute = TRUE,
  cov.compute = FALSE,
  light.return = TRUE,
  bias.correct = FALSE,
  checkNames = FALSE,
  ...
)
}
\arguments{
\item{object}{An \code{KM} object.}

\item{newdata}{Matrix of points where to perform prediction.}

\item{type}{Character giving the kriging type. For now only
\code{"UK"} is possible.}

\item{se.compute}{Logical. Should the standard error be computed?}

\item{cov.compute}{Logical. Should the covariance matrix between
newdata points be computed?}

\item{light.return}{return no other intermediate objects (like T
matrix).}

\item{bias.correct}{fix UK variance and covaariance (defualt is
FALSE)}

\item{checkNames}{check consistency between object design data: X
and newdata (default is FALSE)}

\item{...}{Ignored}
}
\value{
list of predict data: mean, sd, trend, cov, upper95 and
    lower95 quantiles.
}
\description{
Overload DiceKriging::predict.km for \code{KM} objects (expected faster).
}
\examples{
## a 16-points factorial design, and the corresponding response
d <- 2; n <- 16
design.fact <- expand.grid(x1 = seq(0, 1, length = 4), x2 = seq(0, 1, length = 4))
y <- apply(design.fact, 1, DiceKriging::branin) 

## library(DiceKriging)
## kriging model 1 : matern5_2 covariance structure, no trend, no nugget effect
## m1 <- km(design = design.fact, response = y, covtype = "gauss",
##          parinit = c(.5, 1), control = list(trace=F))
as_m1 <- KM(design = design.fact, response = y, covtype = "gauss",
               parinit = c(.5, 1))
as_p <- predict(as_m1,newdata = matrix(.5,ncol = 2), type = "UK",
                checkNames = FALSE, light.return = TRUE)
}
\author{
Yann Richet \email{yann.richet@irsn.fr}
}
