% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{reset_seed}
\alias{reset_seed}
\title{Set random ssed used in fit/optim}
\usage{
reset_seed(x, ...)
}
\arguments{
\item{x}{Seed for mt19937 RNG}

\item{...}{Ignored
#'}
}
\description{
Set random ssed used in fit/optim
}
\author{
Yann Richet \email{yann.richet@irsn.fr}
}
