% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{save}
\alias{save}
\title{Save object.}
\usage{
save(object, ...)
}
\arguments{
\item{object}{An object representing a fitted model.}

\item{...}{Ignored.}
}
\value{
The saved object.
}
\description{
Save a model given in
\code{object}.
}
