% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtils.R
\name{set_num_nugget}
\alias{set_num_nugget}
\title{Setup numerical nugget used in Cholesky decompositions}
\usage{
set_num_nugget(x, ...)
}
\arguments{
\item{x}{Numerical nugget value.}

\item{...}{Ignored}
}
\description{
Setup numerical nugget used in Cholesky decompositions
}
\author{
Yann Richet \email{yann.richet@irsn.fr}
}
