% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asDiceKriging.R
\name{simulate.as_km}
\alias{simulate.as_km}
\alias{simulate,as_km,as_km-method}
\title{Overload DiceKriging::simulate.km for as_km objects (expected faster).}
\usage{
\method{simulate}{as_km}(
  object,
  nsim = 1,
  seed = NULL,
  newdata,
  cond = TRUE,
  nugget.sim = 0,
  checkNames = FALSE,
  ...
)
}
\arguments{
\item{object}{as_km object}

\item{nsim}{number of response vector to simulate}

\item{seed}{random seed}

\item{newdata}{matrix of points where to perform prediction}

\item{cond}{simulate conditional samples (only TRUE accepted)}

\item{nugget.sim}{numercial ngget ,effect to avoid numerical unstabilities}

\item{checkNames}{check consistency between object design data: X and newdata (default is FALSE)}

\item{...}{Ignored}
}
\value{
length(x) x nsim matrix containing simulated path at newdata points
}
\description{
Overload DiceKriging::simulate.km for as_km objects (expected faster).
}
\examples{
f = function(x) 1-1/2*(sin(12*x)/(1+x)+2*cos(7*x)*x^5+0.7)
  plot(f)
set.seed(123)
X <- as.matrix(runif(5))
y <- f(X)
  points(X,y,col='blue')
k <- as_km(design=X, response=y,covtype = "gauss")
x = seq(0,1,,101)
s_x = simulate(k, nsim=3, newdata=x)
  lines(x,s_x[,1],col='blue')
  lines(x,s_x[,2],col='blue')
  lines(x,s_x[,3],col='blue')
}
\author{
Yann Richet (yann.richet@irsn.fr)
}
