% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{simulate.Kriging}
\alias{simulate.Kriging}
\alias{simulate,Kriging,Kriging-method}
\title{Simulate (conditional) Kriging model at given points}
\usage{
\method{simulate}{Kriging}(object, nsim = 1, seed = 123, x, ...)
}
\arguments{
\item{object}{S3 Kriging object}

\item{nsim}{number of simulations to perform}

\item{seed}{random seed used}

\item{x}{points in model input space where to simulate}

\item{...}{Ignored}
}
\value{
length(x) x nsim matrix containing simulated path at x points
}
\description{
Simulate (conditional) Kriging model at given points
}
\examples{
f = function(x) 1-1/2*(sin(12*x)/(1+x)+2*cos(7*x)*x^5+0.7)
  plot(f)
set.seed(123)
X <- as.matrix(runif(5))
y <- f(X)
  points(X,y,col='blue')
r <- Kriging(y, X, "gauss")
x = seq(0,1,,101)
s_x = simulate(r, nsim=3, x=x)
  lines(x,s_x[,1],col='blue')
  lines(x,s_x[,2],col='blue')
  lines(x,s_x[,3],col='blue')
}
\author{
Yann Richet (yann.richet@irsn.fr)
}
