% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoiseKrigingClass.R
\name{simulate.NoiseKriging}
\alias{simulate.NoiseKriging}
\title{Simulation from a \code{NoiseKriging} Object}
\usage{
\method{simulate}{NoiseKriging}(object, nsim = 1, seed = 123, x, ...)
}
\arguments{
\item{object}{S3 NoiseKriging object.}

\item{nsim}{Number of simulations to perform.}

\item{seed}{Random seed used.}

\item{x}{Points in model input space where to simulate.}

\item{...}{Ignored.}
}
\value{
a matrix with \code{length(x)} rows and \code{nsim}
    columns containing the simulated paths at the inputs points
    given in \code{x}.
}
\description{
Simulation from a \code{NoiseKriging} model object.
}
\details{
This method draws paths of the stochastic process at new input
points conditional on the values at the input points used in the
fit.
}
\note{
The names of the formal arguments differ from those of the
    \code{simulate} methods for the S4 classes \code{"km"} and
    \code{"KM"}. The formal \code{x} corresponds to
    \code{newdata}. These names are chosen \pkg{Python} and
    \pkg{Octave} interfaces to \pkg{libKriging}.
}
\examples{
f <- function(x) 1 - 1 / 2 * (sin(12 * x) / (1 + x) + 2 * cos(7 * x) * x^5 + 0.7)
plot(f)
set.seed(123)
X <- as.matrix(runif(5))
y <- f(X) + 0.01*rnorm(nrow(X))
points(X, y, col = "blue")
r <- NoiseKriging(y, X, kernel = "gauss")
x <- seq(from = 0, to = 1, length.out = 101)
s_x <- simulate(r, nsim = 3, x = x)
lines(x, s_x[ , 1], col = "blue")
lines(x, s_x[ , 2], col = "blue")
lines(x, s_x[ , 3], col = "blue")
}
\author{
Yann Richet \email{yann.richet@irsn.fr}
}
