% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asDiceKriging.R
\name{update.as_km}
\alias{update.as_km}
\alias{update,as_km,as_km-method}
\title{Overload DiceKriging::update.km methd for as_km objects (expected faster).}
\usage{
\method{update}{as_km}(
  object,
  newX,
  newy,
  newX.alreadyExist = FALSE,
  cov.reestim = TRUE,
  trend.reestim = cov.reestim,
  nugget.reestim = FALSE,
  newnoise.var = NULL,
  kmcontrol = NULL,
  newF = NULL,
  ...
)
}
\arguments{
\item{object}{as_km object}

\item{newX}{new design points: matrix of object@d columns}

\item{newy}{new response points}

\item{newX.alreadyExist}{if TRUE, newX contains some ppoints already in object@X}

\item{cov.reestim}{fit object to newdata: estimate theta (only supports TRUE)}

\item{trend.reestim}{fit object to newdata: estimate beta (only supports TRUE)}

\item{nugget.reestim}{fit object to newdata: estimate nugget effect (only support FALSE)}

\item{newnoise.var}{add noise to newy response}

\item{kmcontrol}{parametrize fit (unsupported)}

\item{newF}{}

\item{...}{Ignored}
}
\description{
Overload DiceKriging::update.km methd for as_km objects (expected faster).
}
\examples{
f = function(x) 1-1/2*(sin(12*x)/(1+x)+2*cos(7*x)*x^5+0.7)
  plot(f)
set.seed(123)
X <- as.matrix(runif(5))
y <- f(X)
  points(X,y,col='blue')
k <- as_km(design=X, response=y,covtype = "gauss")
x = seq(0,1,,101)
p_x = predict(k, x)
  lines(x,p_x$mean,col='blue')
  lines(x,p_x$lower95,col='blue')
  lines(x,p_x$upper95,col='blue')
newX <- as.matrix(runif(3))
newy <- f(newX)
  points(newX,newy,col='red')
update(k,newy,newX)
x = seq(0,1,,101)
p2_x = predict(k, x)
  lines(x,p2_x$mean,col='red')
  lines(x,p2_x$lower95,col='red')
  lines(x,p2_x$upper95,col='red')
}
\author{
Yann Richet (yann.richet@irsn.fr)
}
