% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asDiceKriging.R
\name{update.KM}
\alias{update.KM}
\alias{update,KM,KM-method}
\title{Overload the `update` Method from \bold{DiceKriging}}
\usage{
\method{update}{KM}(
  object,
  newX,
  newy,
  newX.alreadyExist = FALSE,
  cov.reestim = TRUE,
  trend.reestim = cov.reestim,
  nugget.reestim = FALSE,
  newnoise.var = NULL,
  kmcontrol = NULL,
  newF = NULL,
  ...
)
}
\arguments{
\item{object}{KM object}

\item{newX}{new design points: matrix of object@d columns}

\item{newy}{new response points}

\item{newX.alreadyExist}{if TRUE, newX contains some ppoints
already in object@X}

\item{cov.reestim}{fit object to newdata: estimate theta (only
supports TRUE)}

\item{trend.reestim}{fit object to newdata: estimate beta (only
supports TRUE)}

\item{nugget.reestim}{fit object to newdata: estimate nugget
effect (only support FALSE)}

\item{newnoise.var}{add noise to newy response}

\item{kmcontrol}{parametrize fit (unsupported)}

\item{newF}{New trend matrix.}

\item{...}{Ignored}
}
\description{
Overload DiceKriging::update.km method for `KM` objects (expected faster).
}
\examples{
f <- function(x) 1 - 1 / 2 * (sin(12 * x) / (1 + x) + 2 * cos(7 * x) * x^5 + 0.7)
plot(f)
set.seed(123)
X <- as.matrix(runif(5))
y <- f(X)
points(X,y,col='blue')
k <- KM(design=X, response=y,covtype = "gauss")
x <-  seq(from = 0, to = 1, length.out = 101)
p_x <- predict(k, x)
lines(x, p_x$mean, col = 'blue')
lines(x, p_x$lower95, col = 'blue')
lines(x, p_x$upper95, col = 'blue')
newX <- as.matrix(runif(3))
newy <- f(newX)
points(newX, newy, col = 'red')
update(k, newy, newX)
x <- seq(from = 0, to = 1, length.out = 101)
p2_x <- predict(k, x)
lines(x, p2_x$mean, col = 'red')
lines(x, p2_x$lower95, col = 'red')
lines(x, p2_x$upper95, col = 'red')
}
\author{
Yann Richet \email{yann.richet@irsn.fr}
}
