% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NuggetKrigingClass.R
\name{update.NuggetKriging}
\alias{update.NuggetKriging}
\title{Update a \code{NuggetKriging} Object with New Points}
\usage{
\method{update}{NuggetKriging}(object, newy, newX, ...)
}
\arguments{
\item{object}{S3 NuggetKriging object.}

\item{newy}{Numeric vector of new responses (output).}

\item{newX}{Numeric matrix of new input points.}

\item{...}{Ignored.}
}
\description{
Update a \code{NuggetKriging} model object with new points
}
\section{Caution}{
 The method \emph{does not return the updated
    object}, but instead changes the content of
    \code{object}. This behaviour is quite unusual in R and
    differs from the behaviour of the methods
    \code{\link[DiceKriging]{update.km}} in \pkg{DiceKriging} and
    \code{\link{update,KM-method}}.
}

\examples{
f <- function(x) 1- 1 / 2 * (sin(12 * x) / (1 + x) + 2 * cos(7 * x)*x^5 + 0.7)
plot(f)
set.seed(123)
X <- as.matrix(runif(5))
y <- f(X) + 0.01*rnorm(nrow(X))
points(X, y, col = "blue")
KrigObj <- NuggetKriging(y, X, "gauss")
x <- seq(from = 0, to = 1, length.out = 101)
p_x <- predict(KrigObj, x)
lines(x, p_x$mean, col = "blue")
lines(x, p_x$mean - 2 * p_x$stdev, col = "blue")
lines(x, p_x$mean + 2 * p_x$stdev, col = "blue")
newX <- as.matrix(runif(3))
newy <- f(newX)
points(newX, newy, col = "red")

## change the content of the object 'KrigObj'
update(KrigObj, newy, newX)
x <- seq(from = 0, to = 1, length.out = 101)
p2_x <- predict(KrigObj, x)
lines(x, p2_x$mean, col = "red")
lines(x, p2_x$mean - 2 * p2_x$stdev, col = "red")
lines(x, p2_x$mean + 2 * p2_x$stdev, col = "red")

}
\author{
Yann Richet \email{yann.richet@irsn.fr}
}
