% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{update_simulate}
\alias{update_simulate}
\title{Update simulation of model on data.}
\usage{
update_simulate(object, ...)
}
\arguments{
\item{object}{An object representing a fitted model.}

\item{...}{Further arguments of function}
}
\value{
Updated simulation of model output.
}
\description{
Update previous simulate of a model given in
\code{object}.
}
