% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asDiceKriging.R
\name{as.km.Kriging}
\alias{as.km.Kriging}
\alias{KM,Kriging,Kriging-method}
\title{Coerce a \code{Kriging} Object into the Class \code{"km"}}
\usage{
\method{as.km}{Kriging}(x, .call = NULL)
}
\arguments{
\item{x}{An object with S3 class \code{"Kriging"}.}

\item{.call}{Force the "call" to be filled in \code{km} object.}
}
\value{
An object of having the S4 class \code{"KM"} which
    extends \code{DiceKriging::km} and contains an extra
    \code{"Kriging"} slot.

## @importFrom utils installed.packages XXXY
}
\description{
Coerce a \code{Kriging} object into the class
\code{"km"} from the DiceKriging package.
}
\examples{
f <- function(x) 1 - 1 / 2 * (sin(12 * x) / (1 + x) + 2 * cos(7 * x) * x^5 + 0.7)
set.seed(123)
X <- as.matrix(runif(5))
y <- f(X)
r <- Kriging(y, X, "gauss")
print(r)
k <- KM(r)
print(k)

}
\author{
Yann Richet \email{yann.richet@irsn.fr}
}
