% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.list.Kriging}
\alias{as.list.Kriging}
\alias{as.list,Kriging,Kriging-method}
\title{List Kriging object content}
\usage{
\method{as.list}{Kriging}(x, ...)
}
\arguments{
\item{x}{S3 Kriging object}

\item{...}{Ignored}
}
\value{
list of Kriging object fields: kernel, optim, objective, theta, sigma2, X, centerX, scaleX, y, centerY, scaleY, regmodel, F, T, M, z, beta
}
\description{
List Kriging object content
}
\examples{
f = function(x) 1-1/2*(sin(12*x)/(1+x)+2*cos(7*x)*x^5+0.7)
set.seed(123)
X <- as.matrix(runif(5))
y <- f(X)
r <- Kriging(y, X, "gauss")
l = as.list(r)
cat(paste0(names(l)," =" ,l,collapse="\n"))
}
\author{
Yann Richet (yann.richet@irsn.fr)
}
