% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{leaveOneOut.Kriging}
\alias{leaveOneOut.Kriging}
\alias{leaveOneOut,Kriging,Kriging-method}
\title{Compute leave-One-Out of Kriging model}
\usage{
\method{leaveOneOut}{Kriging}(object, theta, grad = FALSE)
}
\arguments{
\item{object}{S3 Kriging object}

\item{theta}{range parameters to evaluate}

\item{grad}{return Gradient ? (default is TRUE)}
}
\value{
leave-One-Out computed for given theta
}
\description{
Compute leave-One-Out of Kriging model
}
\examples{
f <- function(x) 1 - 1 / 2 * (sin(12 * x) / (1 + x) + 2 * cos(7 * x) * x^5 + 0.7)
set.seed(123)
X <- as.matrix(runif(5))
y <- f(X)
r <- Kriging(y, X, kernel = "gauss", objective = "LOO")
print(r)
loo <-  function(theta) leaveOneOut(r,theta)$leaveOneOut
t <-  seq(from = 0.0001, to = 2, length.out = 101)
plot(t, loo(t), type = 'l')
abline(v = as.list(r)$theta, col = 'blue')
}
\author{
Yann Richet (yann.richet@irsn.fr)
}
