% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{leaveOneOut}
\alias{leaveOneOut}
\title{Compute model leave-One-Out error at given args}
\usage{
leaveOneOut(...)
}
\arguments{
\item{...}{args}
}
\value{
leave-One-Out
}
\description{
Compute model leave-One-Out error at given args
}
