% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLikelihood.Kriging}
\alias{logLikelihood.Kriging}
\alias{logLikelihood,Kriging,Kriging-method}
\title{Compute log-Likelihood of Kriging model}
\usage{
\method{logLikelihood}{Kriging}(object, theta, grad = FALSE, hess = FALSE)
}
\arguments{
\item{object}{S3 Kriging object}

\item{theta}{new points in model output space}

\item{grad}{return Gradient ? (default is TRUE)}

\item{hess}{return Hessian ? (default is FALSe)}
}
\value{
log-Likelihood computed for given theta
}
\description{
Compute log-Likelihood of Kriging model
}
\examples{
f = function(x) 1-1/2*(sin(12*x)/(1+x)+2*cos(7*x)*x^5+0.7)
set.seed(123)
X <- as.matrix(runif(5))
y <- f(X)
r <- Kriging(y, X, "gauss")
print(r)
ll = function(theta) logLikelihood(r,theta)$logLikelihood
t = seq(0.0001,2,,101)
  plot(t,ll(t),type='l')
  abline(v=as.list(r)$theta,col='blue')
}
\author{
Yann Richet (yann.richet@irsn.fr)
}
