% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLikelihood}
\alias{logLikelihood}
\title{Compute model log-Likelihood at given args}
\usage{
logLikelihood(...)
}
\arguments{
\item{...}{args}
}
\value{
log-Likelihood
}
\description{
Compute model log-Likelihood at given args
}
