% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{logLikelihood}
\alias{logLikelihood}
\title{Compute Log-Likelihood}
\usage{
logLikelihood(object, ...)
}
\arguments{
\item{object}{An object representing a fitted model.}

\item{...}{Further arguments for methods.}
}
\value{
The log-likelihood.
}
\description{
Compute the log-Likelihood of a model given in \code{object},
possibly at a different value of the parameters.
}
