% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLikelihood}
\alias{logLikelihood}
\title{Compute Model log-Likelihood at Given Args}
\usage{
logLikelihood(...)
}
\arguments{
\item{...}{args}
}
\value{
log-Likelihood
}
\description{
Compute Model log-Likelihood at Given Args
}
