% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoiseKrigingClass.R
\name{logLikelihoodFun.NoiseKriging}
\alias{logLikelihoodFun.NoiseKriging}
\alias{logLikelihoodFun,NoiseKriging,NoiseKriging-method}
\alias{f}
\alias{<-}
\alias{function(x)}
\alias{1}
\alias{-}
\alias{/}
\alias{2}
\alias{*}
\alias{(sin(12}
\alias{x)}
\alias{(1}
\alias{+}
\alias{cos(7}
\alias{x^5}
\alias{0.7)}
\alias{set.seed(123)}
\alias{X}
\alias{as.matrix(runif(10))}
\alias{y}
\alias{f(X)}
\alias{X/10}
\alias{*rnorm(nrow(X))}
\alias{k}
\alias{NoiseKriging(y,}
\alias{(X/10)^2,}
\alias{X,}
\alias{kernel}
\alias{=}
\alias{"matern3_2")}
\alias{print(k)}
\alias{theta0}
\alias{k$theta()}
\alias{ll_sigma2}
\alias{function(sigma2)}
\alias{logLikelihoodFun(k,}
\alias{cbind(theta0,sigma2))$logLikelihood}
\alias{s2}
\alias{seq(from}
\alias{0.001,}
\alias{to}
\alias{1,}
\alias{length.out}
\alias{101)}
\alias{plot(s2,}
\alias{Vectorize(ll_sigma2)(s2),}
\alias{type}
\alias{'l')}
\alias{abline(v}
\alias{k$sigma2(),}
\alias{col}
\alias{"blue")}
\alias{sigma20}
\alias{k$sigma2()}
\alias{ll_theta}
\alias{function(theta)}
\alias{cbind(theta,sigma20))$logLikelihood}
\alias{t}
\alias{2,}
\alias{plot(t,}
\alias{Vectorize(ll_theta)(t),}
\alias{k$theta(),}
\alias{ll}
\alias{function(theta_sigma2)}
\alias{theta_sigma2)$logLikelihood}
\alias{31)}
\alias{contour(t,s2,matrix(ncol=length(s2),ll(expand.grid(t,s2))),xlab="theta",ylab="sigma2")}
\alias{points(k$theta(),k$sigma2(),col='blue')}
\title{Compute Log-Likelihood of NoiseKriging Model}
\usage{
\method{logLikelihoodFun}{NoiseKriging}(object, theta_sigma2, grad = FALSE, ...)
}
\arguments{
\item{object}{An S3 NoiseKriging object.}

\item{theta_sigma2}{A numeric vector of (positive) range parameters and variance at
which the log-likelihood will be evaluated.}

\item{grad}{Logical. Should the function return the gradient?}

\item{...}{Not used.}
}
\value{
The log-Likelihood computed for given
    \eqn{\boldsymbol{theta_sigma2}}{\theta,\sigma^2}.
}
\description{
Compute Log-Likelihood of NoiseKriging Model
}
\author{
Yann Richet \email{yann.richet@irsn.fr}
}
