% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NuggetKrigingClass.R
\name{logLikelihoodFun.NuggetKriging}
\alias{logLikelihoodFun.NuggetKriging}
\alias{logLikelihoodFun,NuggetKriging,NuggetKriging-method}
\title{Compute Log-Likelihood of NuggetKriging Model}
\usage{
\method{logLikelihoodFun}{NuggetKriging}(object, theta_alpha, grad = FALSE, ...)
}
\arguments{
\item{object}{An S3 NuggetKriging object.}

\item{theta_alpha}{A numeric vector of (positive) range parameters at
which the log-likelihood will be evaluated.}

\item{grad}{Logical. Should the function return the gradient?}

\item{...}{Not used.}
}
\value{
The log-Likelihood computed for given
    \eqn{\boldsymbol{theta_alpha}}{\theta_alpha}.
}
\description{
Compute Log-Likelihood of NuggetKriging Model
}
\examples{
f <- function(x) 1 - 1 / 2 * (sin(12 * x) / (1 + x) + 2 * cos(7 * x) * x^5 + 0.7)
set.seed(123)
X <- as.matrix(runif(5))
y <- f(X) + 0.01*rnorm(nrow(X))
r <- NuggetKriging(y, X, kernel = "gauss")
print(r)
alpha = as.list(r)$sigma2/(as.list(r)$nugget+as.list(r)$sigma2)
ll <- function(theta) logLikelihoodFun(r, c(theta,alpha))$logLikelihood
t <- seq(from = 0.001, to = 2, length.out = 101)
plot(t, ll(t), type = 'l')
abline(v = as.list(r)$theta, col = "blue")

}
\author{
Yann Richet \email{yann.richet@irsn.fr}
}
