% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logMargPost.Kriging}
\alias{logMargPost.Kriging}
\alias{logMargPost,Kriging,Kriging-method}
\title{Compute log-Marginal-Posterior of Kriging model}
\usage{
\method{logMargPost}{Kriging}(object, theta, grad = FALSE)
}
\arguments{
\item{object}{S3 Kriging object}

\item{theta}{range parameters to evaluate}

\item{grad}{return Gradient ? (default is TRUE)}
}
\value{
log-MargPost computed for given theta
}
\description{
Compute log-Marginal-Posterior of Kriging model
}
\examples{
f = function(x) 1-1/2*(sin(12*x)/(1+x)+2*cos(7*x)*x^5+0.7)
set.seed(123)
X <- as.matrix(runif(5))
y <- f(X)
r <- Kriging(y, X, "gauss")
print(r)
lmp <- function(theta) logMargPost(r,theta)$logMargPost
t <- seq(from = 0.0001, to = 2, length.oput = 101)
plot(t, lmp(t), type = 'l')
abline(v = as.list(r)$theta, col = 'blue')
}
\author{
Yann Richet (yann.richet@irsn.fr)
}
