% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NuggetKrigingClass.R
\name{logMargPost.NuggetKriging}
\alias{logMargPost.NuggetKriging}
\alias{logMargPost,NuggetKriging,NuggetKriging-method}
\title{Get logMargPost of NuggetKriging Model}
\usage{
\method{logMargPost}{NuggetKriging}(object, ...)
}
\arguments{
\item{object}{An S3 NuggetKriging object.}

\item{...}{Not used.}
}
\value{
The logMargPost computed for fitted
    \eqn{\boldsymbol{theta}}{\theta}.
}
\description{
Get logMargPost of NuggetKriging Model
}
\examples{
f <- function(x) 1 - 1 / 2 * (sin(12 * x) / (1 + x) + 2 * cos(7 * x) * x^5 + 0.7)
set.seed(123)
X <- as.matrix(runif(5))
y <- f(X)
r <- Kriging(y, X, kernel = "gauss")
print(r)
logMargPost(r)

}
\author{
Yann Richet \email{yann.richet@irsn.fr}
}
