% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{logMargPost}
\alias{logMargPost}
\title{Compute log-Marginal Posterior}
\usage{
logMargPost(object, ...)
}
\arguments{
\item{object}{An object representing a fitted model.}

\item{...}{Further arguments for methods.}
}
\value{
The log-marginal posterior.
}
\description{
Compute the log-Marginal Posterior of a model given in
\code{object}, possibly at a different value of the parameters.
}
