% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logMargPost}
\alias{logMargPost}
\title{Compute model log-Marginal-Posterior at given args}
\usage{
logMargPost(...)
}
\arguments{
\item{...}{args}
}
\value{
log-Marginal-Posterior
 @export
}
\description{
Compute model log-Marginal-Posterior at given args
}
