% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.Kriging}
\alias{predict.Kriging}
\alias{predict,Kriging,Kriging-method}
\title{Predict Kriging model at given points}
\usage{
\method{predict}{Kriging}(object, x, stdev = T, cov = F, ...)
}
\arguments{
\item{object}{S3 Kriging object}

\item{x}{points in model input space where to predict}

\item{stdev}{return also standard deviation (default TRUE)}

\item{cov}{return covariance matrix between x points (default FALSE)}

\item{...}{Ignored}
}
\value{
list containing: mean, stdev, cov
}
\description{
Predict Kriging model at given points
}
\examples{
f = function(x) 1-1/2*(sin(12*x)/(1+x)+2*cos(7*x)*x^5+0.7)
  plot(f)
set.seed(123)
X <- as.matrix(runif(5))
y <- f(X)
  points(X,y,col='blue')
r <- Kriging(y, X, "gauss")
x = seq(0,1,,101)
p_x = predict(r, x)
  lines(x,p_x$mean,col='blue')
  lines(x,p_x$mean-2*p_x$stdev,col='blue')
  lines(x,p_x$mean+2*p_x$stdev,col='blue')
}
\author{
Yann Richet (yann.richet@irsn.fr)
}
