% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrigingClass.R
\name{predict.Kriging}
\alias{predict.Kriging}
\alias{predict,Kriging,Kriging-method}
\title{Predict Kriging Model at Given Points}
\usage{
\method{predict}{Kriging}(object, x, stdev = TRUE, cov = FALSE, ...)
}
\arguments{
\item{object}{S3 Kriging object}

\item{x}{points in model input space where to predict}

\item{stdev}{return also standard deviation (default TRUE)}

\item{cov}{return covariance matrix between x points (default FALSE)}

\item{...}{Ignored}
}
\value{
A list containing: \code{mean}, stdev, cov
}
\description{
Predict Kriging Model at Given Points
}
\examples{
f <- function(x) 1 - 1 / 2 * (sin(12 * x) / (1 + x) + 2 * cos(7 * x) * x^5 + 0.7)
plot(f)
set.seed(123)
X <- as.matrix(runif(5))
y <- f(X)
points(X, y, col = "blue", pch = 16)
r <- Kriging(y, X, "gauss")
x <-seq(from = 0, to = 1, length.out = 101)
p_x <- predict(r, x)
lines(x, p_x$mean, col = "blue")
lines(x, p_x$mean - 2 * p_x$stdev, col = "blue")
lines(x, p_x$mean + 2 * p_x$stdev, col = "blue")
}
\author{
Yann Richet \email{yann.richet@irsn.fr}
}
