% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.Kriging}
\alias{print.Kriging}
\alias{print,Kriging,Kriging-method}
\title{Print Kriging object content}
\usage{
\method{print}{Kriging}(x, ...)
}
\arguments{
\item{x}{S3 Kriging object}

\item{...}{Ignored}
}
\description{
Print Kriging object content
}
\examples{
f = function(x) 1-1/2*(sin(12*x)/(1+x)+2*cos(7*x)*x^5+0.7)
set.seed(123)
X <- as.matrix(runif(5))
y <- f(X)
r <- Kriging(y, X, "gauss")
print(r)
}
\author{
Yann Richet (yann.richet@irsn.fr)
}
