% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{save}
\alias{save}
\title{Save a Kriging Model inside a file. Back to base::save if argument 
is not a Kriging object.}
\usage{
save(object = NULL, filename = NULL, ...)
}
\arguments{
\item{object}{An object representing a model.}

\item{filename}{A file to save the object.}

\item{...}{Arguments used by base::save.}
}
\description{
Save a Kriging Model inside a file. Back to base::save if argument 
is not a Kriging object.
}
\author{
Yann Richet \email{yann.richet@irsn.fr}
}
