% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrigingClass.R
\name{update.Kriging}
\alias{update.Kriging}
\title{Update a \code{Kriging} Object with New Points}
\usage{
\method{update}{Kriging}(object, newy, newX, normalize = FALSE, ...)
}
\arguments{
\item{object}{S3 Kriging object.}

\item{newy}{Numeric vector of new responses (output).}

\item{newX}{Numeric matrix of new input points.}

\item{normalize}{Logical. If \code{TRUE}, the input \code{X} and
the output \code{y} are normalized to lie in \eqn{[0, 1]}.}

\item{...}{Ignored.}
}
\description{
Update a \code{Kriging} model object with new points
}
\examples{
f <- function(x) 1-1/2*(sin(12*x)/(1+x)+2*cos(7*x)*x^5+0.7)
plot(f)
set.seed(123)
X <- as.matrix(runif(5))
y <- f(X)
points(X, y, col = "blue")
r <- Kriging(y, X, "gauss")
x <- seq(from = 0, to = 1, length.out = 101)
p_x <- predict(r, x)
lines(x, p_x$mean, col = "blue")
lines(x, p_x$mean - 2 * p_x$stdev, col = "blue")
lines(x, p_x$mean + 2 * p_x$stdev, col = "blue")
newX <- as.matrix(runif(3))
newy <- f(newX)
points(newX, newy, col = "red")
update(r, newy, newX)
x <- seq(from = 0, to = 1, length.out = 101)
p2_x <- predict(r, x)
lines(x, p2_x$mean, col = "red")
lines(x, p2_x$mean - 2 * p2_x$stdev, col = "red")
lines(x, p2_x$mean + 2 * p2_x$stdev, col = "red")

}
\author{
Yann Richet \email{yann.richet@irsn.fr}
}
